requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
retrying>=1.3.3
urllib3>=1.26.0
certifi>=2021.10.8
pysocks
grobid-client>=0.7.3
json5>=0.9.10
numpy>=1.26.0,<2.0.0
pandas>=1.5.3,<2.0.0
tqdm>=4.65.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0